Add next gen func
